# expence-tracker/my-app/specs/1-expense-tracker-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: RESTful API for the Expense Tracker MVP
  version: 1.0.0
  contact:
    name: API Support
    email: support@expensetracker.example.com
servers:
  - url: https://api.expensetracker.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /categories:
    get:
      tags:
        - Categories
      summary: List categories
      description: Get all predefined and user custom categories
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [EXPENSE, INCOME]
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Categories
      summary: Create custom category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Category name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /categories/{id}:
    put:
      tags:
        - Categories
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot modify predefined category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Categories
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Category deleted
        '400':
          description: Cannot delete category with existing transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Cannot delete predefined category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  # Budgets
...
  /export/csv:
    post:
      tags:
        - Reports
      summary: Export transactions to CSV
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="transactions-2025-11.csv"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [EXPENSE, INCOME]
        isPredefined:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [EXPENSE, INCOME]
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...